+++ app/frontend/components/pages/Arrangements/Index.vue
