+++ app/frontend/components/pages/Profile/Favorites.vue
@@ -1,0 +1,9 @@
+<template>
+  <div>yyyyyyyyyyy</div>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style></style>
