+++ app/frontend/components/parts/buttons/NormalButton.vue
@@ -1,6 +1,17 @@
 <template>
-  <v-btn class="font-weight-bold" depressed outlined :xLarge="xLarge" @click.stop="$emit('click')">
+  <v-btn
+    class="font-weight-bold"
+    depressed
+    outlined
+    :disabled="disabled"
+    :loading="loading"
+    :xLarge="xLarge"
+    @click.stop="$emit('click')"
+  >
     <slot name="text" />
+    <template #loader>
+      <v-progress-circular color="#eeeeee" indeterminate />
+    </template>
   </v-btn>
 </template>
 
@@ -8,6 +19,12 @@
 export default {
   props: {
     xLarge: {
+      type: Boolean,
+    },
+    loading: {
+      type: Boolean,
+    },
+    disabled: {
       type: Boolean,
     },
   },
