+++ app/frontend/components/parts/buttons/SubmitButton.vue
@@ -4,11 +4,15 @@
     depressed
     outlined
     :color="color"
-    :xLarge="xLarge"
     :disabled="disabled"
+    :loading="loading"
+    :xLarge="xLarge"
     @click="$emit('submit')"
   >
     <slot name="text" />
+    <template #loader>
+      <v-progress-circular color="#eeeeee" indeterminate />
+    </template>
   </v-btn>
 </template>
 
@@ -24,6 +28,9 @@ export default {
       default: null,
     },
     disabled: {
+      type: Boolean,
+    },
+    loading: {
       type: Boolean,
     },
   },
